feat: Add Apps module and developer permissions
Introduces a new "Apps" module to the administrative panel, allowing for visualization of installed modules. This change also expands developer permissions to include access to this new module.

Includes updates to type definitions, constants, and routing to accommodate the new module. Additionally, minor refactoring was done to improve express type handling in core and API routes.