feat: Add positioning options for page blocks
Introduces a new `positioning` field to `BlockLayout` to support fixed positioning for page blocks. This allows blocks to be fixed to the top, bottom, left, or right of the viewport, in addition to the existing grid layout.

This change also consolidates `headerBlocks` and `contentBlocks` into a single `mainBlocks` array for simpler management of page content. New icons for fixed positioning have been added to the icon library.